Added option to parse gaia DR3 catalogue to platonium.py
